Added Validation to All GetMapping, to require a session.

+++ src/main/java/now/planit/Controllers/ProjectController.java
@@ -21,6 +21,9 @@ public class ProjectController {
 
     @GetMapping("/myProjects")
     public String myProjects(Model model, WebRequest request) {
+        if (user == null){
+            return "login/login";
+        }
         updateProjects(request, model);
         model.addAttribute("userName", user.getName());
         return "project/myProjects";// endpoint change
@@ -41,6 +44,9 @@ public class ProjectController {
 
     @GetMapping("/removeProject/{id}")
     public String deleteProject(@PathVariable(value = "id") String id, Model model, WebRequest request) {
+        if (user == null){
+            return "login/login";
+        }
         projectService.deleteProject(id, user);
         updateProjects(request,model);
         return "redirect:/myProjects";
