+++ src/main/java/now/planit/Controllers/SubtaskController.java
@@ -23,6 +23,9 @@ public class SubtaskController {
   //Endpoint to pass and connect data(id) from createTask site to createSubtask site.
   @GetMapping("/rerun/{id}")
   public String loadSubtasks(@PathVariable(value = "id") String id, WebRequest request, Model model) {
+    if (user == null){
+      return "login/login";
+    }
     taskName = id;
    updateSubtasks(request, model, taskName);
     return "redirect:/createSubtask";
@@ -31,6 +34,9 @@ public class SubtaskController {
   //Endpoint to dynamic display(loop) all task and display user information
   @GetMapping("/createSubtask")
   public String createSubtasks(WebRequest request, Model model) {
+    if (user == null){
+      return "login/login";
+    }
     updateSubtasks(request, model, taskName);
     return "/project/createSubtask";
   }
@@ -50,6 +56,9 @@ public class SubtaskController {
 
   @GetMapping("/removeSubtask/{id}")
   public String deleteTask(@PathVariable(value = "id") String id, Model model, WebRequest request) {
+    if (user == null){
+      return "login/login";
+    }
     subtaskService.deleteSubtask(taskName, id, user);
     updateSubtasks(request, model, taskName);
     return "redirect:/createSubtask";
